Reorg, fix lint
