+++ blip_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"io"
 	"strings"
 	"testing"
@@ -13,53 +12,6 @@ import (
 	"github.com/localhots/blip"
 	"github.com/localhots/blip/ctx/log"
 )
-
-func TestLogger(t *testing.T) {
-	cfg := blip.DefaultConfig()
-	cfg.Level = blip.LevelDebug
-	log.Setup(cfg)
-	ctx := context.Background()
-	err := errors.New("task already exists")
-
-	log.Debug(ctx, "Message without fields")
-	log.Debug(ctx, "Callback received", log.F{
-		"device_unique_id": "G4000E-1000-F",
-		"task_id":          123456,
-		"status":           "success",
-		"template_name":    "index.tpl",
-	})
-	log.Info(ctx, "Starting task", log.F{
-		"device_unique_id": "G4000E-1000-F",
-		"task_id":          123456,
-	})
-	log.Info(ctx, "Extremely long message, sorry about that; it is meant to prove that buffer can grow"+strings.Repeat(" @@@", 300), log.F{
-		"device_unique_id": "G4000E-1000-F",
-		"task_id":          123456,
-	})
-	log.Warn(ctx, "Duplicate task, but also this message exceeds 40 characters", log.F{
-		"task_id": 123456,
-	})
-	log.Warn(ctx, "Duplicate task but exactly 40 characters", log.F{
-		"task_id": 123456,
-	})
-	log.Warn(blip.WithContext(ctx, log.F{"foo": "bar"}), "Duplicate task is exactly 39 characters", log.F{
-		"task_id": 123456,
-	})
-	log.Error(ctx, "Failed to process task", log.Cause(err), log.F{
-		"task_id": 123456,
-	})
-}
-
-func TestFatal(t *testing.T) {
-	cfg := blip.DefaultConfig()
-	log.Setup(cfg)
-	ctx := context.Background()
-	err := errors.New("task already exists")
-
-	log.Fatal(ctx, "Failed to start service", log.Cause(err), log.F{
-		"service": "api",
-	})
-}
 
 //
 // Fuzz
