+++ encoder_json_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"testing"
 )
 
@@ -54,25 +53,6 @@ func TestJSONEncoderNoFieldsNoTime(t *testing.T) {
 	validateAndPrintJSON(t, buf)
 }
 
-func TestJSONEncoderMulti(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.Level = LevelDebug
-	cfg.Encoder = NewJSONEncoder(cfg)
-
-	logger := New(cfg)
-	ctx := context.Background()
-
-	logger.Info(ctx, "Starting task")
-	logger.Info(ctx, "Starting task", F{
-		"device_unique_id": "G4000E-1000-F",
-		"task_id":          123456,
-	})
-	logger.Info(ctx, "Starting task", F{
-		"device_unique_id": "G4000E-1000-F",
-		"task_id":          123456,
-	})
-}
-
 func validateAndPrintJSON(t *testing.T, buf bytes.Buffer) {
 	t.Helper()
 	if buf.Len() == 0 {
@@ -83,5 +63,4 @@ func validateAndPrintJSON(t *testing.T, buf bytes.Buffer) {
 	if err := json.Unmarshal(buf.Bytes(), &data); err != nil {
 		t.Errorf("Failed to unmarshal JSON: %v", err)
 	}
-	fmt.Print(buf.String())
 }
